Added Text class
Basic animated text class which will print out a string at a given rate.
Next will be to assign any one text entity to a character so they can
say their lines, and add restrictions to the length and wrapping of the
line based on position and screen width. Maybe make it return a bool for
when it is and isn't updating so it can block other actions.