Speech is now position relative
M command key added for testing purposes, shifts Matt along by 30
pixels. His speech will now render a fixed offset from his coordinates.

Next I will make a function that renders the text on the side of him
that is closest to the center, this will require length limiting and
multi line text rendering. To the right it will not need to be
positioned based on length but to the left it will so short lines don't
render too far away.