Wrote basic template for text offset function
Text offset function to decide which side of speaking sprite to render
text on etc, not yet implemented and half of the inputs are not
generated yet - still compiles because the method is not called.

Added Global class for dealing with stuff which I'll probably have to
set a little more dynamically later on - works well to keep tabs on what
is hard coded in however, will add all hard-coded data in here so it can
be whittled down later.

Also added in N as Matt#posx -= 30 for testing purposes.